I'm so damn indecisive! Ergh
I don’t want it to look so boxy, I intend to make it better I just want
to get all the code in there and then mess with it to look the way I
want it to. Just hoping this isn’t like my actual artistic ability…I
have a big idea and then it never turns out the way I thought! :D YAY